+++ interp/src/interpreter/mod.rs
@@ -1,8 +1,9 @@
 mod interpret_component;
 mod interpret_control;
 mod interpret_group;
-mod simulation_utils;
 mod steppers;
+mod utils;
 mod working_environment;
 
 pub use interpret_component::interpret_component;
+pub use steppers::{ComponentInterpreter, Interpreter};
