+++ interp/src/interpreter/steppers/mod.rs
@@ -2,4 +2,6 @@ mod component_interpreter;
 mod control_interpreter;
 mod group_interpreter;
 
+pub use component_interpreter::ComponentInterpreter;
+pub use control_interpreter::Interpreter;
 pub use group_interpreter::AssignmentInterpreter;
